Add AbstractRepositoryFactory
