+++ lib/Doctrine/ODM/MongoDB/Repository/AbstractRepositoryFactory.php
@@ -1,0 +1,83 @@
+<?php
+/*
+ * THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS
+ * "AS IS" AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT
+ * LIMITED TO, THE IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR
+ * A PARTICULAR PURPOSE ARE DISCLAIMED. IN NO EVENT SHALL THE COPYRIGHT
+ * OWNER OR CONTRIBUTORS BE LIABLE FOR ANY DIRECT, INDIRECT, INCIDENTAL,
+ * SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES (INCLUDING, BUT NOT
+ * LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES; LOSS OF USE,
+ * DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND ON ANY
+ * THEORY OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT
+ * (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE
+ * OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
+ *
+ * This software consists of voluntary contributions made by many individuals
+ * and is licensed under the MIT license. For more information, see
+ * <http://www.doctrine-project.org>.
+ */
+
+namespace Doctrine\ODM\MongoDB\Repository;
+
+use Doctrine\ODM\MongoDB\DocumentManager;
+use Doctrine\ODM\MongoDB\Mapping\ClassMetadata;
+
+/**
+ * Abstract factory for creating document repositories.
+ *
+ * @since 1.2
+ */
+abstract class AbstractRepositoryFactory implements RepositoryFactory
+{
+    /**
+     * The list of DocumentRepository instances.
+     *
+     * @var array<\Doctrine\Common\Persistence\ObjectRepository>
+     */
+    private $repositoryList = array();
+
+    /**
+     * {@inheritdoc}
+     */
+    public function getRepository(DocumentManager $documentManager, $documentName)
+    {
+        $metadata = $documentManager->getClassMetadata($documentName);
+        $hashKey = $metadata->getName() . spl_object_hash($documentManager);
+
+        if (isset($this->repositoryList[$hashKey])) {
+            return $this->repositoryList[$hashKey];
+        }
+
+        $repository = $this->createRepository($documentManager, ltrim($documentName, '\\'));
+
+        $this->repositoryList[$hashKey] = $repository;
+
+        return $repository;
+    }
+
+    /**
+     * Create a new repository instance for a document class.
+     *
+     * @param DocumentManager $documentManager The DocumentManager instance.
+     * @param string          $documentName    The name of the document.
+     *
+     * @return \Doctrine\Common\Persistence\ObjectRepository
+     */
+    protected function createRepository(DocumentManager $documentManager, $documentName)
+    {
+        $metadata            = $documentManager->getClassMetadata($documentName);
+        $repositoryClassName = $metadata->customRepositoryClassName ?: $documentManager->getConfiguration()->getDefaultRepositoryClassName();
+
+        return $this->instantiateRepository($repositoryClassName, $documentManager, $metadata);
+    }
+
+    /**
+     * Instantiates requested repository.
+     *
+     * @param string $repositoryClassName
+     * @param DocumentManager $documentManager
+     * @param ClassMetadata $metadata
+     * @return \Doctrine\Common\Persistence\ObjectRepository
+     */
+    abstract protected function instantiateRepository($repositoryClassName, DocumentManager $documentManager, ClassMetadata $metadata);
+}
