+++ lib/Doctrine/ODM/MongoDB/Repository/DefaultRepositoryFactory.php
@@ -20,56 +20,18 @@
 namespace Doctrine\ODM\MongoDB\Repository;
 
 use Doctrine\ODM\MongoDB\DocumentManager;
+use Doctrine\ODM\MongoDB\Mapping\ClassMetadata;
 
 /**
  * This factory is used to create default repository objects for entities at runtime.
  */
-class DefaultRepositoryFactory implements RepositoryFactory
+class DefaultRepositoryFactory extends AbstractRepositoryFactory
 {
-    /**
-     * The list of DocumentRepository instances.
-     *
-     * @var array<\Doctrine\Common\Persistence\ObjectRepository>
-     */
-    private $repositoryList = array();
-
     /**
      * {@inheritdoc}
      */
-    public function getRepository(DocumentManager $documentManager, $documentName)
+    protected function instantiateRepository($repositoryClassName, DocumentManager $documentManager, ClassMetadata $metadata)
     {
-        $metadata = $documentManager->getClassMetadata($documentName);
-        $hashKey = $metadata->getName() . spl_object_hash($documentManager);
-
-        if (isset($this->repositoryList[$hashKey])) {
-            return $this->repositoryList[$hashKey];
-        }
-
-        $repository = $this->createRepository($documentManager, ltrim($documentName, '\\'));
-
-        $this->repositoryList[$hashKey] = $repository;
-
-        return $repository;
-    }
-
-    /**
-     * Create a new repository instance for a document class.
-     *
-     * @param DocumentManager $documentManager The DocumentManager instance.
-     * @param string          $documentName    The name of the document.
-     *
-     * @return \Doctrine\Common\Persistence\ObjectRepository
-     */
-    protected function createRepository(DocumentManager $documentManager, $documentName)
-    {
-        $metadata            = $documentManager->getClassMetadata($documentName);
-        $repositoryClassName = $metadata->customRepositoryClassName;
-
-        if ($repositoryClassName === null) {
-            $configuration       = $documentManager->getConfiguration();
-            $repositoryClassName = $configuration->getDefaultRepositoryClassName();
-        }
-
         return new $repositoryClassName($documentManager, $documentManager->getUnitOfWork(), $metadata);
     }
 }
